SC-353 Handle Invalid State (#141)
* SC-353: DNS Utility

inspect() returns 'subverted' or 'reverted' based on stdout from
dns_utility command being 127.0.0.1 or not.

* SC-353: DNS Utility

inspect() returns 'subverted' or 'reverted' based on stdout from
dns_utility command being 127.0.0.1 or not.

* SC-353: Invalid logic and styling

Adding invalid logic to node_actuator and styled it.

* SC-353: Refator process -> getStatus

This method name sounded more appropriate and with the pending
changes to the underlying DNS utility of a status option it was
time to refactor

* SC-353: Return boolean instead of class list obj

If we find the class in the list return true otherwise false

* SC-353: Subvert/Revert methods WIP

Adding subvert and revert methods to dns_utility. Flushing out the
remaining changes to node_actuator will likley drive more changes
here as well so, WIP.

* SC-353: Fixing renderer_test

Adding node-status-buttons and verification to the bind call

* SC-353: Lint changes

After doing a full build the lint process changed a few things

* SC-353: Refactor to us dns_utility status

Moving to status from inspect as this is more succinct

* SC-353: Status Handler

Moving the logic for handling status in the UI to an EventEmitter
as we will need to be able to control the status from multiple
places.

* SC-353: Handle invalid state

Set state to invalid when revert errors.

* SC-353: Refactor node_actuator

Use dns_utility for subverting and reverting DNS. Extracted status
logic to the status_handler

* SC-353: Spike Nodejs find-process

* SC-353: init-status

Adding a handler for init-status to initialize the status of the UI
by asking the status of DNS and the node.

* SC-353: Kill existing SubstratumNode on stopNode

Find the SubstratumNode process by name and kill it on stopNode if
the UI did not start the child process.

* SC-353: Lint changes

yarn lint changed some things...

* SC-353: Init status on load.

Removed spike code and replaced with a call to the status_handler
to initialize the status on load.

* SC-353: Moving to ps-list

find-process did not work for all platforms so moved to ps-lint

* SC-353: Removing console log

Removed a console log that was not needed.

* SC-353: Rename find_wrapper -> ps_wrapper

After moving from find-process to ps-list nodejs module it was
time to rename this class.

* SC-353: Refactor findWrapper -> psWrapper

After moving from find-process to ps-list refactoring variable
names to reflect.

* SC-353: State could be off|serving|consuming

Check that a button is active on startup. This probably should be
more sophiticated but I am lacking imagination right now.