SC-392: Find Node process include path (#145)
* SC-392: Find Node process include path

refactorted findByName and killByName to findNodeProcess and
killNodeProcess. These methods now look at the command path of
the process to determine whether or not it is the actual Node.
The UI is packaged as a binary and that process was also named
SubstratumNode, which caused the invalid state from SC-353 to
break.

* SC-392: Replace ps-list

ps-list uses tasklist for windows and does not return the command
path. Brought ps-list in and replaced tasklist with wmic in order
to get the command path.

Performance side benefit as tasklist was horribly slow.

* SC-392: PID parseInt

Need to make the PID a number for Windows

* SC-392: Split tests

Splitting the tests into thier own tests for better output

* SC-392: More lenient

Changed finds itself test to be more lenient.

* SC-392: Change terms & services to conditions

The documentation is Terms & Conditions

* SC-392: Linting

Commit linting changes

* SC-392: Revert called twice

DNS utility revert was called twice.

* SC-392: Linting changes

Committing changes from the linter

* SC-392: Bring status back into node_actuator

The status_handler was not able to be fully utlized from command
processes and for now I am bringing back into node actuator.

* SC-392: Fix node_actualor.nodeIsRunning

* SC-392: Remove status_handler

No longer used, removing

* SC-392: Increasing main window height

Windows was displaying a scroll bar after packaging release.

* SC-392: Fixed start node

Start node was getting the right value whether the node was
running or not.

* SC-392: Adding ps-list copyright

MIT license requires use to include the copyright since we
copied a large portion of their code. It was modified to use
wmic on Windows instead of tasklist.