SC-728 Report Accounts Payable Request - Exit (#257)
* SC-728: sending message to accountant for requested exit services

* SC-728: small cleanup

* SC-728: multinode integration test failing at unimplemented

* SC-728: multinode integration test passing, but could be improved upon

* SC-728: Adding the payload bytes to log.

Added the number of payload bytes to the log message for the
accountant. Refactored SubstratumNode.wrote_log_containg to accept
a pattern and Regex the logs for that pattern. Refactored accountant
exit service multi-node integration test to be ambiguous on which
exit node the Accountant logs for.

* SC-728: Code cleanup.

Minor refactoring for readability.

* SC-728: PR comments issues.

Changed the word found to requested for debug log statement and
refactored tests to match.