SC-742: Avoid undesirable exit nodes. (#335)
* SC-742: Avoid undesirable exit nodes.

In make_route_segment we now sort possible routes exit nodes
by their desirable flag as we take the first one from the list
for the route. If the target is Some in make_route_segment
then we will not sort the possible routes as the exit nodes
will all be the target. This allows us to avoid undesirable exit
nodes but not exclude them entirely.

* SC-742: Sort routes by desirable exit nodes handle unexpected input.

When we are passed empty routes, route segments, or if we cannot find
a NodeRecord in our database by one of the supplied PublicKeys,
crash the entire application.